backend/requirements.txt
Django
djangorestframework
django-cors-headers

# pip install -r requirements.txt
